- need to stop the previous lists from resetting to top scroll tho

Merge branch 'master' of https://github.com/SupSuper/OpenXcom into oXc_kL
Conflicts:
	bin/data/Language/en-GB.yml
	bin/data/Language/en-US.yml
	src/Basescape/CraftEquipmentState.cpp
	src/Basescape/CraftEquipmentState.h
	src/Basescape/CraftSoldiersState.cpp
	src/Basescape/CraftSoldiersState.h
	src/Basescape/NewManufactureListState.cpp
	src/Basescape/SoldiersState.cpp
	src/Battlescape/AbortMissionState.cpp
	src/Battlescape/ActionMenuState.cpp
	src/Battlescape/AlienBAIState.cpp
	src/Battlescape/BattlescapeGame.cpp
	src/Battlescape/BattlescapeGenerator.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/InfoboxOKState.cpp
	src/Battlescape/InfoboxState.cpp
	src/Battlescape/MedikitState.cpp
	src/Battlescape/MedikitView.cpp
	src/Battlescape/MiniMapState.cpp
	src/Battlescape/MiniMapView.cpp
	src/Battlescape/NextTurnState.cpp
	src/Battlescape/NextTurnState.h
	src/Battlescape/PrimeGrenadeState.cpp
	src/Battlescape/ScannerState.cpp
	src/Battlescape/UnitSprite.cpp
	src/Engine/Game.cpp
	src/Engine/Options.inc.h
	src/Engine/Palette.cpp
	src/Engine/Zoom.cpp
	src/Geoscape/DogfightState.cpp
	src/Geoscape/Globe.cpp
	src/Interface/NumberText.cpp
	src/Interface/TextList.cpp
	src/Interface/TextList.h
	src/Menu/ListGamesState.h
	src/Menu/OptionsAdvancedState.cpp
	src/Menu/OptionsAudioState.cpp
	src/Menu/OptionsControlsState.cpp
	src/Menu/OptionsVideoState.cpp
	src/Menu/PauseState.cpp
	src/Menu/SaveGameState.cpp
	src/Ruleset/RuleCraft.cpp
	src/Ruleset/RuleCraft.h
	src/Ruleset/RuleTerrain.cpp
	src/Ruleset/RuleTerrain.h
	src/Ruleset/Ruleset.cpp
	src/Savegame/AlienMission.cpp
	src/Savegame/Base.cpp
	src/Savegame/Craft.cpp
	src/Savegame/SavedBattleGame.cpp
	src/Savegame/SavedBattleGame.h
	src/Ufopaedia/ArticleStateTFTD.cpp

+++ src/Basescape/SoldiersState.cpp
@@ -229,7 +229,7 @@ void SoldiersState::btnArmorClick(Action*)
 {
 	_game->pushState(new CraftArmorState(
 										_base,
-										static_cast<size_t>(0)));
+										0));
 }
 
 /**
@@ -266,7 +266,7 @@ void SoldiersState::lstSoldiersClick(Action* action)
  */
 void SoldiersState::lstLeftArrowClick(Action* action) // kL
 {
-	int row = _lstSoldiers->getSelectedRow();
+	size_t row = _lstSoldiers->getSelectedRow();
 	if (row > 0)
 	{
 		Soldier* soldier = _base->getSoldiers()->at(row);
@@ -276,7 +276,7 @@ void SoldiersState::lstLeftArrowClick(Action* action) // kL
 			_base->getSoldiers()->at(row) = _base->getSoldiers()->at(row - 1);
 			_base->getSoldiers()->at(row - 1) = soldier;
 
-			if (row != static_cast<int>(_lstSoldiers->getScroll()))
+			if (row != _lstSoldiers->getScroll())
 			{
 				SDL_WarpMouse(
 						static_cast<Uint16>(action->getLeftBlackBand() + action->getXMouse()),
@@ -303,11 +303,11 @@ void SoldiersState::lstLeftArrowClick(Action* action) // kL
  */
 void SoldiersState::lstRightArrowClick(Action* action) // kL
 {
-	int row = _lstSoldiers->getSelectedRow();
+	size_t row = _lstSoldiers->getSelectedRow();
 	size_t numSoldiers = _base->getSoldiers()->size();
 	if (numSoldiers > 0
 		&& numSoldiers <= INT_MAX
-		&& row < static_cast<int>(numSoldiers) - 1)
+		&& row < numSoldiers - 1)
 	{
 		Soldier* soldier = _base->getSoldiers()->at(row);
 
@@ -316,7 +316,7 @@ void SoldiersState::lstRightArrowClick(Action* action) // kL
 			_base->getSoldiers()->at(row) = _base->getSoldiers()->at(row + 1);
 			_base->getSoldiers()->at(row + 1) = soldier;
 
-			if (row != static_cast<int>(_lstSoldiers->getVisibleRows() - 1 + _lstSoldiers->getScroll()))
+			if (row != _lstSoldiers->getVisibleRows() - 1 + _lstSoldiers->getScroll())
 			{
 				SDL_WarpMouse(
 						static_cast<Uint16>(action->getLeftBlackBand() + action->getXMouse()),
